bosh.js: Don't respond to a stream add request before the server accepts the connection. The exception is for the first stream on a BOSH session. Additionally, do not respond with a stream add response if this is the first stream on this BOSH session since it is not required.
xmpp-proxy-connector.js: Add comments and handle the on-connected event. Also, buffer unsent data in case it arrives before the connection to the XMPP Server (by the proxy) is established.

xmpp-proxy.js: Get rid of _in_error in lieu if _is_connected.

There is some sort of abstraction leakage where xmpp-proxy-connector.js knows about the private member of XMPPProxy. This needs to be fixed.